+++ src/App.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Login from './pages/login/login';
+import Login from './pages/loginPage/login';
 import { Switch, Route } from 'react-router-dom';
 import './App.scss';
 import PrivateRoute from './utils/privateRoute';
